Added tag functionality. Imported songs to db.
Also made the database correctly link to the full path. This is a hacky
solution, but it'll have to do for now. Eventually I'll move the
database to a suitable long-term location.

Also imported all the songs into the database, and a few tags + links to
get me started.

I want to support searching for multiple tags in a later version, and I
also want to be able to add/remove tags, and add/remove links.