Added motion detection, lag detection
This is currently slightly broken on MY cameras, just because I've configured
them to have a timestamp at the top-left. Thus, the changing timestamp causes
false alarms for motion detection.

I have not tested the lag detection at all. There's nothing wrong with my code,
but it doesn't run well on slow hardware (e.g: raspberry pi). This might help
detect slowness. If that does happen, I might have to throttle the framerate or
something.

Moved out a bunch of the motion detection and image manipulation functions to
a util package. They have overlap with some of the other projects I've done
before (such as the image simplification).

There's still lots of debugging stuff here, and FIXME's to make this more
generic, especially for high-framerate hardware. But it's a damn good start.