Added support for nightvision and multiprocess saving
Changed to using kernel to dilate with, for circular pattern, instead of square.

Added support for nightvision, with changing sensitivity
Added is_greyscale function for checking if the camera is using nightvision.

Added some starter code for a multi-process implementation that saves frames.
With this, we can tell another process to save the video/images, while the
main process focusses on listening to the stream.

The multi-process stuff is VERY early, and not doing anything at the moment.
It's all debug, really. I might want to swap from using Pipe to Queue. It
feels like multiprocessing.Pipe is REALLY clunky, and doesn't work very well.

Made motion "clump" together a little more by increasing values. Motion still
doesn't actually do anything, but it does now display in the debug output!